backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
drf-jwt==1.19.2
environ==1.0
filelock==3.8.0
idna==3.4
joblib==1.1.0
numpy==1.21.2
oauthlib==3.2.1
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.0
sklearn==0.0
sqlparse==0.4.3
threadpoolctl==3.1.0
types-cryptography==3.3.23
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
